+++ RELEASES.md
@@ -1,3 +1,8 @@
+# 0.5.3
+* Fixed bug when calcualting the enteperise field.
+* Cleanup ipfix code.
+* Rust 2024 Edition.
+
 # 0.5.2
 * Can now parse enterprise fields in non options templates for IPFIX.